+++ packages/cubejs-testing-drivers/test/databricks-jdbc-export-bucket-full.test.ts
@@ -1,0 +1,3 @@
+import { testQueries } from '../src/tests/testQueries';
+
+testQueries('databricks-jdbc', { includeIncrementalSchemaSuite: true, extendedEnv: 'export-bucket' });
