+++ packages/cubejs-testing-drivers/test/databricks-jdbc-full.test.ts
@@ -1,4 +1,3 @@
 import { testQueries } from '../src/tests/testQueries';
 
 testQueries('databricks-jdbc', { includeIncrementalSchemaSuite: true });
-testQueries('databricks-jdbc', { includeIncrementalSchemaSuite: true, extendedEnv: 'export-bucket' });
